cmd copy: move annotations funcs to annotations.py
